+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.7.6-dev
  - properly subtract previous statistic when handling `set_failure()` and `set_success()`
+ - detect and track redirects in `GooseRawRequest`
 
 ## 0.7.5 June 10, 2020
  - store actual URL requested in GooseRawRequest